requirements.txt
pytest==7.2.2
einops==0.6.0
wheel==0.40.0
lion-pytorch==0.0.6
scipy==1.11.4
